Update Azure Function App deployment process
Refactor `vfr3d-cron-functions-centralus-test.yml` to enhance the build and deployment workflow. Removed the setup for a separate .NET Core version for in-process extensions. Added steps to create a `.azurefunctions` directory and package the deployment into a zip file. Updated the deployment method to use Azure CLI instead of the Azure Functions Action, with added verification steps for deployment status.